requirements.txt
django==6.0.2
django-cors-headers
python-dotenv
google-api-python-client==2.191.0
google-auth-httplib2==0.3.0
google-auth-oauthlib==1.3.0

# RAG / semantic search
chromadb        # includes built-in ONNX embedding (all-MiniLM-L6-v2, no PyTorch needed)
python-docx

